abstract `Platform` into it's own file
